requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
blinker==1.9.0
build==1.3.0
CacheControl==0.14.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cleo==2.1.0
click==8.3.1
crashtest==0.4.1
cryptography==46.0.3
dash==3.3.0
dash_mantine_components==2.4.0
distlib==0.4.0
dnspython==2.8.0
dulwich==0.24.10
email-validator==2.3.0
fastapi==0.123.5
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.2
fastar==0.8.0
fastjsonschema==2.21.2
filelock==3.20.0
findpython==0.7.1
Flask==3.1.2
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jeepney==0.9.0
Jinja2==3.1.6
keyring==25.7.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
msgpack==1.1.2
narwhals==2.13.0
nest-asyncio==1.6.0
packaging==25.0
pbs-installer==2025.11.20
pkginfo==1.12.1.2
platformdirs==4.5.0
plotly==6.5.0
poetry==2.2.1
poetry-core==2.2.1
psycopg==3.3.1
psycopg-binary==3.3.1
pwdlib==0.3.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyproject_hooks==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
RapidFuzz==3.14.3
requests==2.32.5
requests-toolbelt==1.0.0
retrying==1.4.2
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
SecretStorage==3.5.0
sentry-sdk==2.46.0
setuptools==80.9.0
shellingham==1.5.4
SQLAlchemy==2.0.44
starlette==0.50.0
tomlkit==0.13.3
trove-classifiers==2025.12.1.14
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
virtualenv==20.35.4
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.4
zipp==3.23.0
zstandard==0.25.0
